Make the OrderFieldMigration available for third-party developers
